Changed player script and rename old weapon.
I changed the player script slightly so that it can work for the multiple weapons. I also renamed the current weapon in the scene to snake bite revolver since that is what it is. I also mapped the secondary fire button.
I also added some new folders to be used for better organizing in the prefab and script folders.