Throwing and modifications
Added a throwing script that can be used on weapons so that when the circle button is clicked the weapon will spin away from the player. Weapon will also not be able to be used until the throwing is over. Need to add more for ammo and time limit considerations. I also modified the player script a little to add second fire and throw buttons. The snake revolver script was also changed for the throwing mechanic.