Camera movement, fix throwing, fix head
I added camera movement that will calculate player distance and expand the camera field of view depending on the distance between players. I also fixed the bugs for throwing and added a stopping mechanic for when it hits a player. I also fixed the player head so that it faces the correct direction. I also changed some of the snake revolver script to make it nicer for future weapons.