Added map, fixed bugs, collisions, Golden knife
Added a whitebox map in the test scene to test. Fixed bugs on the camera so the growing is now based on the center of the map (used for small map). Added collisions for both bullets and thrown weapons so that they both stop when hitting a wall in the map. I also added Greed's default weapon. Added all characters into test scene (Emerald not done yet).