Event manager and enemies
Previous commit the big map camera was also added (forgot to mention). Worked on the event manager and the enemy spawner. Got it so that enemies will increase in damage and health each time they are spawned. Made it so that the enemies can be damaged by the player’s default weapons.