Added Eater and walk/run for player.
Added a walking and running mechanic on the player that is dependent on how far the left stick is being pushed. Added eater script and added it to the event manager so that it will spawn after a certain period as a harbinger. Also made it become a miniboss after a certain number of enemy spawns.