enemy colors, eater health, Cooldown/ammo slider
Made it so that the enemies can now change colors when damaged and when poisoned. Also added a eater health bar above it that will go down when eater is damaged. Also added a placeholder cooldown/ammo slider above the player that goes down when using a weapon or up when player is using the default weapon to show the cooldown.