Lava move up, change lost spawning, bug fixing, Hud changes, sakura map smaller
I made it so that the lava will now move upwards. Changed the way the lost spawn into the game so that they spawn nicer and more spread out. Fixed some bugs with animation and weapons. Changed the way the Hud chooses players for the hud to now find players instead of it being given. Also made the sakura map smaller.