Fixing enemy bugs, Eater aura particle, enemy water Particles
Fixed some bugs with the enemies not showing their hit and poisons colors correctly. Added an Eater aura particle effect to make it seem more menacing and to better tell players that it is a boss. Added the enemy water splash particles to show players that they are being slowed.